requirements.txt
numpy
pip
pypng
setuptools